requirements.txt
asgiref==3.3.4
Django==3.2.2
djangorestframework==3.12.4
greenlet==1.1.0
msgpack==1.0.2
mysqlclient==2.0.3
pynvim==0.4.3
pytz==2021.1
sqlparse==0.4.1
